Factor mutation related stuff into dedicated file
