+++ website/src/utils/mutation.ts
@@ -1,0 +1,131 @@
+import type { BaseType } from './sequenceTypeHelpers';
+import type { ReferenceGenomesSequenceNames } from '../types/referencesGenomes';
+
+export type MutationQuery = {
+    baseType: BaseType;
+    mutationType: 'substitutionOrDeletion' | 'insertion';
+    text: string;
+};
+
+export const parseMutationString = (
+    value: string,
+    referenceGenomesSequenceNames: ReferenceGenomesSequenceNames,
+): MutationQuery[] => {
+    return value
+        .split(',')
+        .map((mutation) => {
+            const trimmedMutation = mutation.trim();
+            if (isValidNucleotideMutationQuery(trimmedMutation, referenceGenomesSequenceNames)) {
+                return { baseType: 'nucleotide', mutationType: 'substitutionOrDeletion', text: trimmedMutation };
+            }
+            if (isValidAminoAcidMutationQuery(trimmedMutation, referenceGenomesSequenceNames)) {
+                return { baseType: 'aminoAcid', mutationType: 'substitutionOrDeletion', text: trimmedMutation };
+            }
+            if (isValidNucleotideInsertionQuery(trimmedMutation, referenceGenomesSequenceNames)) {
+                return { baseType: 'nucleotide', mutationType: 'insertion', text: trimmedMutation };
+            }
+            if (isValidAminoAcidInsertionQuery(trimmedMutation, referenceGenomesSequenceNames)) {
+                return { baseType: 'aminoAcid', mutationType: 'insertion', text: trimmedMutation };
+            }
+            return null;
+        })
+        .filter(Boolean) as MutationQuery[];
+};
+
+export const serializeMutationQueries = (selectedOptions: MutationQuery[]): string => {
+    return selectedOptions.map((option) => option.text).join(', ');
+};
+
+export const isValidAminoAcidInsertionQuery = (
+    text: string,
+    referenceGenomesSequenceNames: ReferenceGenomesSequenceNames,
+): boolean => {
+    try {
+        const textUpper = text.toUpperCase();
+        if (!textUpper.startsWith('INS_')) {
+            return false;
+        }
+        const query = textUpper.slice(4);
+        const [gene, position, insertion] = query.split(':');
+        const existingGenes = new Set(referenceGenomesSequenceNames.genes.map((g) => g.toUpperCase()));
+        if (!existingGenes.has(gene) || !Number.isInteger(Number(position))) {
+            return false;
+        }
+        return /^[A-Z*?]+$/.test(insertion);
+    } catch (_) {
+        return false;
+    }
+};
+
+export const isValidAminoAcidMutationQuery = (
+    text: string,
+    referenceGenomesSequenceNames: ReferenceGenomesSequenceNames,
+): boolean => {
+    try {
+        const textUpper = text.toUpperCase();
+        const [gene, mutation] = textUpper.split(':');
+        const existingGenes = new Set(referenceGenomesSequenceNames.genes.map((g) => g.toUpperCase()));
+        if (!existingGenes.has(gene)) {
+            return false;
+        }
+        return /^[A-Z*]?[0-9]+[A-Z-*\\.]?$/.test(mutation);
+    } catch (_) {
+        return false;
+    }
+};
+
+export const isValidNucleotideInsertionQuery = (
+    text: string,
+    referenceGenomesSequenceNames: ReferenceGenomesSequenceNames,
+): boolean => {
+    try {
+        const isMultiSegmented = referenceGenomesSequenceNames.nucleotideSequences.length > 1;
+        const textUpper = text.toUpperCase();
+        if (!textUpper.startsWith('INS_')) {
+            return false;
+        }
+        const query = textUpper.slice(4);
+        const split = query.split(':');
+        const [segment, position, insertion] = isMultiSegmented
+            ? split
+            : ([undefined, ...split] as [undefined | string, string, string]);
+        if (segment !== undefined) {
+            const existingSegments = new Set(
+                referenceGenomesSequenceNames.nucleotideSequences.map((n) => n.toUpperCase()),
+            );
+            if (!existingSegments.has(segment)) {
+                return false;
+            }
+        }
+        if (!Number.isInteger(Number(position))) {
+            return false;
+        }
+        return /^[A-Z*?]+$/.test(insertion);
+    } catch (_) {
+        return false;
+    }
+};
+
+export const isValidNucleotideMutationQuery = (
+    text: string,
+    referenceGenomesSequenceNames: ReferenceGenomesSequenceNames,
+): boolean => {
+    try {
+        const isMultiSegmented = referenceGenomesSequenceNames.nucleotideSequences.length > 1;
+        const textUpper = text.toUpperCase();
+        let mutation = textUpper;
+        if (isMultiSegmented) {
+            const [segment, _mutation] = textUpper.split(':');
+            const existingSegments = new Set(
+                referenceGenomesSequenceNames.nucleotideSequences.map((n) => n.toUpperCase()),
+            );
+            if (!existingSegments.has(segment)) {
+                return false;
+            }
+            mutation = _mutation;
+        }
+        return /^[A-Z]?[0-9]+[A-Z-\\.]?$/.test(mutation);
+    } catch (_) {
+        return false;
+    }
+};
